Add desktop navbar
Examples of doing the navbar "the right way" (aka not the convuluted
way I have done it for internships in school) for mobile were not
working. For now I will just use the nav on desktop, but eventually
the nav will be responsive to hide the desktop nav items and show
mobile sidebar. Most likely using angular's flex-module.