Add archived line statuses, require fields in form group
A LOT happened in this commit and TBH I had troubles thinking
how to break this up. Adding the archive made me feel as if
form groups were neccisary for validation, so this all flowed
together into one cohesive change. I'll break this explanation
up into two parts: Archiving and Form Group

---ARCHIVING---

I renamed the `LineStatus` interface to `LineState` to
represent the current state of the line that the supervisor is
setting. A new interface `LineStatus` relates to all four fields
of a line's status, like `date` or `comment`. I get now that the
field name and the component name and the interface all called
"Line Status" is still very confusing and I will rename it later.

The supervisor sets all of the fields in the form to a
`lineStatusInstance`. Once clicking submit, all the fields of the
instance are saved into the `lineStatusArray`and that overall
"line status" with all the fields are the newest line status in
the archive. `lineStatusInstance` fields are all reset to empty
strings.

---FORM GROUP---

To ensure that all fields but `comment` were required, they were
all grouped in a `FormGroup` called `lineStatusForm`. Each field
in the form is given a `FormControl` definition, which registers
the field to the `lineStatusForm`.

To click the submit button to add a line status to the archive,
all `Form Control` fields in the `lineStatusForm` `FormGroup`
except comment must be filled out. The button will be disabled
otherwise. Once enabled, the click event on the button,
`submitLineStatus()` pushes an instance of the line status to
the archive and clears all fields. The form is reset, clearing
all form inputs.

---CONCLUSION---

I am sorry for the vague naming as of this point, and the large
size of the commit. I hope this expanation helped.

Addresses #9 (Add previous dates to line status)