Add `json-server` for temporary dev DB, fetch `users` from auth
[`json-server`](https://www.npmjs.com/package/json-server?activeTab=versions)
 is being used temporarily for a development DB until the choice
of backend is made. It can be rum witj `json-server --watch db.json`.
Two `user`s were added to the `users` array. For reference,
this as a backend being used for Angular2+ came from this
[SitePoint article](https://www.sitepoint.com/angular-rxjs-create-api-service-rest-backend/)

`AuthService` grabs the users with an observable, so users should
be able to be accessed in any component neccisary. In this case,
`LineStatusComponent` needs to know a user's role. And to test
different roles, `NavbarCompoment` has a toggle that will soon be
able to switch between the roles `user` and `admin`.

`user`, for consideration as of now, is the default role of a user
which has minimal access. Eventually, users will be stored in the
real DB with whatever backend is used, and tied to whatever
authentication service is used. Firebase, a likely candidate, can
handle both the DB and authentication.