Remote redundant fields from LineStatusComponent
Line of thinking with the previous code:
I set `lineStatusInstance` with all fields initialized to empty
strings, and then fields with the same name initialized empty
outside of the instance. The idea was that the standalone fields
would be less verbose to read in the template and would be their
own data until `submitLineStatus()` would capture them to be the
respective fields of `lineStatusInstance`.

Cause for change:
However, this made `submitLineStatus()` more verbose with
unneccisary assignment.

Future changes:
The next step that could have logically been in this commit would
be to shorten `lineStatusInstance` to just `lineStatus`. However,
I feel like I could achieve the effect of what I wanted in the
first paragraph. Simple field names in the template. I feel like
ES6 object destructering could achieve a similar thing, as I have
seen it do very cool things in React. However, I can't translate
that to reality at the moment despite skimming some guides, so
that will have to wait. The fact that I throw it out as a buzzword
with no explanation should be evidence enough. Yet I just know it
is the best candidate for the solution.

Addresses PR #6 (Create "Line Status" page)