Sort line statuses shallowly by past/present/future in ngOnInit
Statuses that have been set as of component initialization, aka
the hardcoded ones, are filtered from the `lineStatusArray` into
past/present/future LineStatus arrays by date. However, new
entries are not sorted into those categories as of form
submission, so they don't show up in the view. That is out of
scope for this commit.

In a similair sort of vein, the hardcoded statuses were listed
in an order that would look as if the system knows how to order
shifts in a "proper" chronological order. As in:
* Filtering doesn't sort days within a category, like Dec 1 of
the past and Dec 2 of the past.
* Filtering doesn't assume some order of night>day>mid.
* Filtering doesn't order statuses in a single shift on a single
day, like trial running @ noon and then cleaning at 1pm.

Some of those examples are loaded with assumptions that need to
be discussed with the customers, like what the order of shifts
are in a "day". IE is a day structured traditionally like
day>mid>night, or practically night>day>mid like many days workout
as I assumed. It is because of these assumptions that no logic is
set in place as of yet until customer feedback, though the
seed data will be ordered in such a way based on assumed behaivor.
If it seems easy to impliment and trash, some code may sneak its
way in.

Addresses PR #6 (Create "Line Status" page)