Place future and archived status cards in a `mat-expansion-panel`
Future and past archives will add up and are not as important to
the average user, but should be viewable if they look. Both
sections are now in `mat-expansion-panel` areas to shrink down the
page size unless the user wants to see more.

The styling is all over the place though. I made some haphazard
changes to make it bearable. However, some flaws in the structure
are evident. Like blanket styles on elements, like <section>, and
the varying "center" classes.

Addresses PR #6 (Create "Line Status" page)